+++ src/loaders.js
@@ -4,19 +4,10 @@ import Page from './page.js';
  * Class to load an url and generate a page with the result
  */
 export class UrlLoader {
-    constructor(url, options) {
+    constructor(url, options = {}) {
         this.url = url;
         this.options = options;
         this.html = null;
-    }
-
-    /**
-     * Performs a fetch to the url and return a promise
-     *
-     * @return {Promise}
-     */
-    fetch() {
-        return fetch(this.url, this.options);
     }
 
     /**
@@ -37,7 +28,7 @@ export class UrlLoader {
             return new Promise((accept) => accept(new Page(this.url, parseHtml(this.html))));
         }
 
-        return this.fetch()
+        return fetch(this.url, this.options)
             .then((res) => {
                 if (res.status < 200 || res.status >= 300) {
                     throw new Error(`The request status code is ${res.status}`);
@@ -62,19 +53,30 @@ export class UrlLoader {
  * Class to submit a form and generate a page with the result
  */
 export class FormLoader extends UrlLoader {
-    constructor(form, options) {
-        let url = form.action.split('?', 2).shift();
-        const method = (form.method || 'GET').toUpperCase();
+    constructor(form, options = {}) {
+        if (!options.method) {
+            options.method = (form.method || 'GET').toUpperCase();
+        }
 
-        if (method === 'GET') {
+        let url;
+
+        if (options.url) {
+            url = options.url;
+            delete options.url;
+        } else {
+            url = form.action.split('?', 2).shift();
+        }
+
+        if (options.method === 'GET') {
             url += '?' + new URLSearchParams(new FormData(form));
+        } else if (options.method === 'POST' && !options.body) {
+            options.body = new FormData(this.form);
         }
 
         super(url, options);
 
+        //Disable cache
         this.html = false;
-        this.method = method;
-        this.form = form;
     }
 
     /**
@@ -82,25 +84,6 @@ export class FormLoader extends UrlLoader {
      */
     fallback() {
         this.form.submit();
-    }
-
-    /**
-     * Performs a fetch with the form data and return a promise
-     *
-     * @return {Promise}
-     */
-    fetch() {
-        const options = { method: this.method };
-
-        if (this.method === 'POST') {
-            options.body = new FormData(this.form);
-        }
-
-        if (this.options) {
-            Object.assign(options, this.options);
-        }
-
-        return fetch(this.url, options);
     }
 }